Changes and fixes resulting from review. Changed type of stop reason from long to DWORD. FetchRemoteServerList now in has try block around HTTPSRequest. Now only getting SessionID once in GetStatusRequest. No longer setting STOP_REASON_ALL as a signal anywhere, and added a comment to indicate that it shouldn't be done.

+++ src/connectionmanager.h
@@ -45,7 +45,7 @@ public:
 
     ServerList& GetServerList();
     void Toggle(const tstring& transport, bool startSplitTunnel);
-    void Stop(long reason);
+    void Stop(DWORD reason);
     void Start(const tstring& transport, bool startSplitTunnel);
     void StartSplitTunnel();
     void StopSplitTunnel();
